Extract registry to its own module
Ensure it can be reused and the primitives of std.concurrency do not depend on it.
The `unregisterMe` function has been removed,
as it was a cleanup function for when a Task finished.
If this kind of integration is to exist, which we might not want to,
it needs to happen at the Scheduler level and the registry needs to be configurable.