Remove code duplication between LocalRest and concurrency
This code duplication was historical, because issues with std.concurrency were discovered.
Nowadays that we don't use std.concurrency but have the module in LocalRest,
we can greatly simplify this code.
Note that as the removed unittest shows, we lose the ability to manually synchronize
fibers, as conditions are not used anymore between Fibers.
This will be solved by the revamped concurrency framework, and is currently not needed.