Remove setMailboxSize and associated
As we are aiming to implement concurrency with a more Go-style API, those primitive become unnecessary.
The API offers the ability to change the size of a mailbox, however in practice it is seldom used,
and can be easily replaced.
The more important distinction to be made is between buffered (asynchronous) channels,
and unbuffered (synchronous) channels.
The former offer no hand-off mechanism (meaning writing to such a channel might not block),
while the latter will always result in a context switch (which requires the Scheduler's cooperation).
In this framework, the ability to set the mailbox size stands as an odd primitive,
as it only accounts for buffered channels. It is also an unnecessary source of complexity.