Fix issue 64: Responding to a dead peer makes LocalRest crash
The root of the issue was that, while trying to ensure better channel semantics
(notably, not silently ignoring put on closed channel), we uncovered a bug,
where responding to a dead peer would lead to the responder to crash.
This was visible in Agora, where there are cross-nodes scheduled polling operations,
and the nodes are shut down one after the other at the end of a test.

Using `trySend` here ensures that whatever happens, we don't fail our node.