Report internal error when a node dies
Currently, if there is any bug in LocalRest implementation,
a node can die silently (the thread just stops).
That can lead to pretty hard to identify bugs.
Instead of silently loosing a thread, we catch `Throwable`
and at least print a message before re-throwing
(which will most likely lead to a thread shutdown).