Introduce geod24.concurrency.trySend to account for closed channels
As can be seen in issue 64, it is possible that a node gets a request,
and by the time it is done handling it and ready to respond,
the requester is dead. In this situation, we don't want our node
to die (it's essentially equivalent to an RST in TCP parlance),
only to drop the request.
In order to do this, we need the ability to do "optional" send.