Call a custom callback routine on ctrl.shutdown()
Previously, if the Node had any shutdown behavior
it wishes to implement but where this behavior
could *not* be implemented in its destructor
then the client code would need to implement a
custom shutdown() API endpoint.

However, there is a race condition:

After the call to node.shutdown(),
and before the call to node.ctrl.shutdown(),
another thread (another node) may have a pending
request into this same node. And this call will
succeed, because node.ctrl.shutdown() was not
called yet and this handler thread was not
destroyed yet.

The 'shutdown' method in the control interface
can now optionally be provided a function
to be called just before the node is destroyed.

It takes as its parameter the API,
so the function may call introspect or call
the Node's methods before that Node is destroyed.