Scheduler: FiberCondition need not inherit from Condition
A Condition should be tied to a Mutex. The general idea is that one acquire the Mutex,
checks a 'condition' (e.g. x <= 0), and call 'wait' with or without a timeout.
The wait effectively releases the lock, and when another Thread completes the condition,
it will call 'notify' or 'notifyAll'.
Users of 'Condition' usually need to loop on the condition being checked,
as multiple listeners might be competing on the same 'Condition'
(hence why there is notify and notifyAll).
However, a FiberCondition does nothing of the sort. What a FiberCondition boils down
to is a repeated condition check and yielding, which in the multi-threaded world
is more akin to a spinlock than an actual, cooperative 'Condition',
because the Scheduler will re-schedule Fibers which are waiting for a condition to complete,
even before 'Condition.notify[All]' has been called.