Make all tests use thread_joinAll
While not required (tests are independent and so are nodes),
it is more hygienic to do so at the end of the test.
In addition, if a node lives on by mistake, it'll be easier
to track down (because we'll block on the test and get a stack trace).
The final advantage is that it makes things easier to read in debuggers,
as only the threads used for the current tests are visible.