Make client (request) error a separate type of error
Some client logic might rely on treating errors differently,
e.g. HTTP has 4xx and 5xx error codes,
where 4xx could be used to drive the application logic.
While we don't expose errors the same way as HTTP,
we have a use case for having a separate kind of `Exception` for client errors,
the first of which would be a syntactically invalid query.