Add a scheduler ID to avoid range errors
If a node had a timed-out request and then
was restarted via the ctrl interface,
this request may still receive a response
from the target node. However at that point
the LocalScheduler instance was cleared,
and the request map is missing the previous
ID of the request.

Adding a scheduler ID makes it easy
to filter out these outdated responses.

A test-case was added, which would ordinarly
cause a range exception without this bugfix.