Refactor Registry to prepare for 'bind connections'
Currently, the registry only contains Tid, which can be used to send and receive messages.
As we want to move from the typeless Tid to a typed approach (using channels),
we need to abstract away some details. In that process, documentation was
marginally improved and some style issues were addressed.