Strongly type and differentiate the listening connection to the remote node
Since our framework follow the same model as the basic TCP/IP server,
we have two kinds of connections: the "main" connection, or listener,
which in C would be the file descriptor given by `bind(2)`.
Whenever a client wants to connect to us, they reach us through that
connection, and we establish a new client connection.
In C, that would be the file descriptor returned by `accept`,
in our framework it is a `RemoteAPI`.