Fix blocking in Ctor edge case
Since #116, to allow starting tasks in Ctor of nodes,
LocalRest registers the FiberScheduler before starting
the dispatch loop. This might cause the main thread to try
on blocking Fiber level primitives, which is not possible.

isMainThread() should check for a Fiber context rather than
a running FiberScheduler instance and runTask() should be happy
with a scheduler instance even if it is running in the main thread.