Fix self conn deadlock
When a node goes to sleep and manages to send
a request to itself the request gets queued. Once
it wakes up from the sleep it will try to process
that request synchronous to own event loop and locks
itself. Queued messages should be handled async to the
event loop.