Registry: Add clear() method
When a network shuts down, it usually needs to unregister all its nodes first,
in order to avoid tasks from accidentally acquiring new clients and creating
random bugs (usually SEGV).

However, clearing the whole `Registry` is expensive:
Each successful `unregister` call will do two lookups in the connection AA,
one in the names array, along with a remove call. And that is on top of the
lock acquision / release for each `unregister` call.

Doing two `clear` calls on the AA is much cheaper.